Added Enemy Collision & Damage Cooldown
Added enemy collision (red rectangle easy level enemy). Also made it so there's a cooldown between damage taken by player so that player doesn't die immediately when touching enemy. Health is printed (decreases appropriately), not yet shown on screen though.

+++ gameproject/src/main/java/com/game/Character.java
@@ -3,7 +3,7 @@ package com.game;
 public abstract class Character {
     public int currentHealth;
     public int maxHealth;
-    public int moveSpeed;
+    public double moveSpeed;
     public int power;
     public boolean isDead;
 }
